backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9; python_version < "3.14"
psycopg[binary]==3.3.3; python_version >= "3.14"
